massive rework of files and extrapolation of components etc
