+++ src/components/Carousel/HorizontalSwipeCarousel.tsx
@@ -1,7 +1,7 @@
 import { STORE_IMAGE_SIZES } from 'constants/config'
 import { useLimitedHorizontalSwipe } from 'hooks/useScrollAnimation'
-import AnimatedCarousel from './AnimatedCarousel'
-import { BaseCarouselProps, WithTouchAction } from './types'
+import AnimatedCarousel from './common/components/AnimatedCarousel'
+import { BaseCarouselProps, WithTouchAction } from './common/types'
 
 export default function HorizontalSwipeCarousel({
   data,
