+++ src/components/Carousel/LargeProductImageCarousel.tsx
@@ -1,4 +1,4 @@
-import LargeImageCarouselModal from 'components/LargeImageCarouselModal'
+import LargeImageCarouselModal from 'components/Carousel/common/components/LargeImageModalCarousel'
 import SmartImg, { SmartImageProps } from 'components/SmartImg'
 import { STORE_IMAGE_SIZES } from 'constants/config'
 import { GenericImageSrcSet } from 'shopify/graphql/types'
