+++ src/components/Carousel/common/components/styleds.ts
