+++ src/components/Carousel/common/hooks.ts
