+++ src/pages/App.tsx
@@ -6,7 +6,7 @@ const Popups = lazy(() => import(/* webpackPrefetch: true,  webpackChunkName: "P
 const Collection = lazy(() => import(/* webpackPrefetch: true,  webpackChunkName: "COLLECTION" */ 'pages/Collection'))
 const NotFound = lazy(() => import(/* webpackChunkName: "NOTFOUND" */ 'pages/Error/NotFound'))
 const Navigation = lazy(() => import(/* webpackChunkName: "NAVIGATION" */ 'components/Navigation'))
-const SingleItem = lazy(() => import(/* webpackChunkName: "SINGLEITEM" */ 'pages/SingleItem'))
+const SingleItem = lazy(() => import(/* webpackChunkName: "SINGLEITEM" */ 'pages/SingleProduct'))
 
 import { FallbackLoader } from 'components/Loader'
 import { useQuery } from '@apollo/client'
